requirements.txt
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
crc==6.1.1
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.2
flask-redis==0.4.0
Flask-SocketIO==5.3.6
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-redis==0.4.0
python-socketio==5.11.2
redis==5.0.3
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
